fix(nerf): move alpha compositing logic to CUDA kernel, details below:
The alpha compositing method used prior to this commit worked correctly
for input RGBA images, but _not_ for volume-rendered pixels.  For
volume-rendered pixels, the accumulated opacity along a ray has already
been factored in the composited color of this ray, thus calling the
`blend_alpha_channel` function would multiply the alpha (ray's opacity)
value another time.

This commit does not modify the `blend_alpha_channel` function but moves
the background compositing logic to the CUDA kernel for differentiable
volume rendering.

It might be worth noting that although the `final_rgbs` is now comprised
of the accumulated colors along the ray and optionally a background
color, the calculation of gradients from `dL_dfinal_rgbs` to
`ray_dL_ddensities` does not change.

Signed-off-by: Gaoyang Zhang <gy@blurgy.xyz>