chore: workaround a possible JAX regression
The commit that (possibly) introduced the regression: <google/jax#12611>

It does not yet affect the current jaxngp codebase (because our JAX
version does not contain the commit that introduced this regression),
but does affect the jaxngp codebase if the nixpkgs flake is bumped to
the 23.05 tag, where the commit has been packaged to nixpkgs while the
revert <google/jax#12852> has not.

This commit thus works as a future-proof, and also makes the custom
differentiation implementation consistent between the two custom CUDA
extensions `jax-tcnn` and `volume-rendering-jax`.

Signed-off-by: Gaoyang Zhang <gy@blurgy.xyz>